[App Search] Minor log retention refactors (#88140)
* Move reused log retention components to own folder

- e.g., logic file, helpers used by other views & not just settings

* Update settings-specific views to reference top-level folder

- I'm leaving the panel and the confirmation modal within settings as that isn't used anywhere outside the settings page, but we can revisit this if needed or if people think it makes more sense to keep everything log related together

* [bug] Fix nested <p> error in log retention confirmation modal
