+++ x-pack/plugins/enterprise_search/public/applications/app_search/components/log_retention/messaging/index.test.tsx
@@ -4,8 +4,8 @@
  * you may not use this file except in compliance with the Elastic License.
  */
 
-import '../../../../../__mocks__/kea.mock';
-import { setMockValues } from '../../../../../__mocks__';
+import '../../../../__mocks__/kea.mock';
+import { setMockValues } from '../../../../__mocks__';
 
 import React from 'react';
 import { shallow } from 'enzyme';
