+++ x-pack/plugins/enterprise_search/public/applications/app_search/components/log_retention/types.ts
