+++ x-pack/plugins/enterprise_search/public/applications/app_search/components/log_retention/utils/convert_log_retention.test.ts
@@ -3,6 +3,7 @@
  * or more contributor license agreements. Licensed under the Elastic License;
  * you may not use this file except in compliance with the Elastic License.
  */
+
 import { convertLogRetentionFromServerToClient } from './convert_log_retention';
 
 describe('convertLogRetentionFromServerToClient', () => {
