+++ x-pack/plugins/enterprise_search/public/applications/app_search/components/settings/log_retention/generic_confirmation_modal.tsx
@@ -55,7 +55,7 @@ export const GenericConfirmationModal: React.FC<GenericConfirmationModalProps> =
           <p>
             <strong>{subheading}</strong>
           </p>
-          <p>{description}</p>
+          {description}
         </EuiText>
         <EuiSpacer />
         <EuiFormRow
