+++ x-pack/plugins/enterprise_search/public/applications/app_search/components/settings/log_retention/index.ts
@@ -1,0 +1,8 @@
+/*
+ * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
+ * or more contributor license agreements. Licensed under the Elastic License;
+ * you may not use this file except in compliance with the Elastic License.
+ */
+
+export { LogRetentionPanel } from './log_retention_panel';
+export { LogRetentionConfirmationModal } from './log_retention_confirmation_modal';
